Separate public and solver-related utilities in separate folders
