+++ dynamiqs/solvers/ode/adaptive_solver.py
@@ -8,10 +8,10 @@ import torch
 from torch import Tensor
 from tqdm.std import TqdmWarning
 
-from ...utils.progress_bar import tqdm
-from ...utils.solver_utils import hairer_norm
 from ...utils.tensor_types import dtype_complex_to_real
 from ..solver import AutogradSolver
+from ..utils.progress_bar import tqdm
+from ..utils.solver_utils import hairer_norm
 
 
 class AdaptiveSolver(AutogradSolver):
