+++ dynamiqs/solvers/ode/fixed_solver.py
@@ -7,9 +7,9 @@ import torch.nn as nn
 from torch import Tensor
 from torch.autograd.function import FunctionCtx
 
-from ...utils.progress_bar import tqdm
-from ...utils.solver_utils import add_tuples, none_to_zeros_like
 from ..solver import AdjointSolver, AutogradSolver
+from ..utils.progress_bar import tqdm
+from ..utils.solver_utils import add_tuples, none_to_zeros_like
 
 
 class FixedSolver(AutogradSolver):
