+++ dynamiqs/solvers/propagator.py
@@ -2,9 +2,9 @@ from abc import abstractmethod
 
 from torch import Tensor
 
-from ..utils.progress_bar import tqdm
-from ..utils.td_tensor import ConstantTDTensor
 from .solver import AutogradSolver
+from .utils.progress_bar import tqdm
+from .utils.td_tensor import ConstantTDTensor
 
 
 class Propagator(AutogradSolver):
