+++ dynamiqs/solvers/utils/progress_bar.py
